Fixed package name for the search task, renamed message for search by limit and offset task
