+++ CoreFeatures/Database-postgresql/PostgresSearchByLimitAndOffsetTask/src/main/java/info/smart_tools/smartactors/database_postgresql/postgres_search_by_limit_and_offset_task/PostgresSearchByLimitAndOffsetTask.java
@@ -109,7 +109,7 @@ public class PostgresSearchByLimitAndOffsetTask implements IDatabaseTask {
     @Override
     public void prepare(final IObject query) throws TaskPrepareException {
         try {
-            SearchMessage message = IOC.resolve(Keys.getKeyByName(SearchMessage.class.getCanonicalName()), query);
+            SearchByLimitAndOffsetMessage message = IOC.resolve(Keys.getKeyByName(SearchByLimitAndOffsetMessage.class.getCanonicalName()), query);
             collection = message.getCollectionName();
             criteria = message.getCriteria();
             callback = message.getCallback();
