+++ CoreFeatures/Database-postgresql/PostgresSearchByLimitAndOffsetTask/src/main/java/info/smart_tools/smartactors/database_postgresql/postgres_search_by_limit_and_offset_task/SearchByLimitAndOffsetMessage.java
@@ -8,7 +8,7 @@ import info.smart_tools.smartactors.iobject.iobject.exception.ReadValueException
 /**
  * A query message to be passed to {@link PostgresSearchByLimitAndOffsetTask#prepare(IObject)}.
  */
-public interface SearchMessage {
+public interface SearchByLimitAndOffsetMessage {
 
     /**
      * Returns the collection name where to search the document
