--- CoreFeatures/Database-postgresql/PostgresSearchTask/src/main/java/info/smart_tools/smartactors/database_postgresql/postgres_search_by_limit_and_offset_task/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Database task to search the document by it's fields from PostgreSQL database.
- */
-package info.smart_tools.smartactors.database_postgresql.postgres_search_by_limit_and_offset_task;
