+++ CoreFeatures/Database-postgresql/PostgresSearchTask/src/main/java/info/smart_tools/smartactors/database_postgresql/postgres_search_task/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Database task to search the document by it's fields from PostgreSQL database.
+ */
+package info.smart_tools.smartactors.database_postgresql.postgres_search_task;
